# clients/openapi-v2.json
{
 "components": {
  "schemas": {
...
   "EpisodePatch": {
    "additionalProperties": false,
    "properties": {
     "base_prompt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Base Prompt"
     },
     "order": {
      "anyOf": [
       {
        "minimum": 0.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Order"
     },
     "revision": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "description": "Optimistic concurrency check.",
      "title": "Revision"
     },
     "strip": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Strip-Input"
       },
       {
        "type": "null"
       }
      ]
     },
     "synopsis": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Synopsis"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "title": "EpisodePatch",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "PanelIds": {
    "properties": {
     "panel_ids": {
      "items": {
       "type": "string"
      },
      "title": "Panel Ids",
      "type": "array"
     }
    },
    "required": [
     "panel_ids"
    ],
    "title": "PanelIds",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "bearer": {
    "bearerFormat": "mio_<id>_<secret>",
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "description": "Open API of Mio Comic Studio. Authenticate with `Authorization: Bearer mio_…` (create tokens in Settings → API). Scopes: read, write, render, admin.",
  "title": "Mio API",
  "version": "2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/episodes/{episode_id}": {
   "delete": {
    "operationId": "trash_episode",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Trash Episode",
    "tags": [
     "series"
    ]
   },
   "get": {
    "operationId": "get_episode",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Get Episode",
    "tags": [
     "series"
    ]
   },
   "patch": {
    "operationId": "patch_episode",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EpisodePatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Patch Episode",
    "tags": [
     "series"
    ]
   }
  },
...
  "/episodes/{episode_id}/panels": {
   "post": {
    "operationId": "add_panel",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PanelCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Add Panel",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/batch": {
   "post": {
    "description": "Batch edit: one revision, every listed panel gets the same changes.",
    "operationId": "patch_panels",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PanelBatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Patch Panels",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/batch-delete": {
   "post": {
    "operationId": "delete_panels",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PanelIds"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Delete Panels",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/import": {
   "post": {
    "operationId": "import_panels",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PanelImport"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Import Panels",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/reorder": {
   "post": {
    "operationId": "reorder_panels",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Reorder"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Reorder Panels",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/{panel_id}": {
   "delete": {
    "description": "Removes the panel; its takes stay in the episode (orphaned takes are hidden, not lost).",
    "operationId": "delete_panel",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "panel_id",
      "required": true,
      "schema": {
       "title": "Panel Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Delete Panel",
    "tags": [
     "series"
    ]
   },
   "patch": {
    "operationId": "patch_panel",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "panel_id",
      "required": true,
      "schema": {
       "title": "Panel Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PanelPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Patch Panel",
    "tags": [
     "series"
    ]
   }
  },
...
  "/episodes/{episode_id}/panels/{panel_id}/duplicate": {
   "post": {
    "operationId": "duplicate_panel",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "panel_id",
      "required": true,
      "schema": {
       "title": "Panel Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Duplicate Panel",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/{panel_id}/history": {
   "get": {
    "description": "Earlier saved versions of one panel (newest first, unchanged saves collapsed).",
    "operationId": "panel_history",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "panel_id",
      "required": true,
      "schema": {
       "title": "Panel Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response Panel History",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Panel History",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/{panel_id}/prompt": {
   "get": {
    "description": "Compiled prompt in both dialects plus the automatically chosen reference images.",
    "operationId": "panel_prompt",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "panel_id",
      "required": true,
      "schema": {
       "title": "Panel Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "variant_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Variant Id"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Panel Prompt",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Panel Prompt",
    "tags": [
     "series"
    ]
   }
  },
  "/episodes/{episode_id}/panels/{panel_id}/restore": {
   "post": {
    "description": "Put the panel back to how it was at ``revision`` (position and id stay).",
    "operationId": "restore_panel",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "panel_id",
      "required": true,
      "schema": {
       "title": "Panel Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PanelRestore"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Episode"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Restore Panel",
    "tags": [
     "series"
    ]
   }
  },
...
  "/episodes/{episode_id}/restore": {
   "post": {
    "operationId": "restore_episode",
    "parameters": [
     {
      "in": "path",
      "name": "episode_id",
      "required": true,
      "schema": {
       "title": "Episode Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Restore Episode",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Restore Episode",
    "tags": [
     "series"
    ]
   }
  },
...
 },
 "servers": [
  {
   "url": "/api/v2"
  }
 ]
}